+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Minor Improvements
 
 * Add edit SVG to Icon
+* Supports Ajax call for error validation
 
 # 0.16.0