+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 * Add edit SVG to Icon
 * Supports Ajax call for error validation
 
+# 0.16.1
+
+* Add reload function to browser helper
+
 # 0.16.0
 
 ## Minor Improvements